Fix Blueprint: No Markdown in wpcli-markdown  (#34)
![Screenshot 2024-04-21 at 08 32
18](https://github.com/adamziel/blueprints/assets/39980/cae5578a-bfa7-46fa-973a-677fc9cea2c1)
As there is no markdown in the content file anymore after I switched it
to block notation, I had to

- change the name,  
- change the description
- rename the directory, 
- update the references to new directory. 

Background: Originally, it was supposed to be markdown. But testing
revealed the import of the markdown into blocks didn't work. The post
was rendered as classic block and markdown notation was ignored. So I
switched to creating the content from Block Markup.